# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oficina Mecânica API",
    "version": "1.0.0",
    "description": "MVP — Sistema Integrado de Atendimento"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/clientes/": {
      "post": {
        "summary": "Criar cliente",
        "tags": [
          "Clientes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "cpfCnpj": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "telefone": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "endereco": {
                    "type": "string",
                    "maxLength": 500
                  }
                },
                "required": [
                  "nome",
                  "cpfCnpj"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "cpfCnpj": {
                      "type": "string"
                    },
                    "tipoPessoa": {
                      "type": "string",
                      "enum": [
                        "FISICA",
                        "JURIDICA"
                      ]
                    },
                    "email": {
                      "type": "string",
                      "nullable": true
                    },
                    "telefone": {
                      "type": "string",
                      "nullable": true
                    },
                    "endereco": {
                      "type": "string",
                      "nullable": true
                    },
                    "ativo": {
                      "type": "boolean"
                    },
                    "criadoEm": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "atualizadoEm": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "nome",
                    "cpfCnpj",
                    "tipoPessoa",
                    "ativo",
                    "criadoEm",
                    "atualizadoEm"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar clientes",
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "busca",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "nome": {
                            "type": "string"
                          },
                          "cpfCnpj": {
                            "type": "string"
                          },
                          "tipoPessoa": {
                            "type": "string",
                            "enum": [
                              "FISICA",
                              "JURIDICA"
                            ]
                          },
                          "email": {
                            "type": "string",
                            "nullable": true
                          },
                          "telefone": {
                            "type": "string",
                            "nullable": true
                          },
                          "endereco": {
                            "type": "string",
                            "nullable": true
                          },
                          "ativo": {
                            "type": "boolean"
                          },
                          "criadoEm": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "atualizadoEm": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "nome",
                          "cpfCnpj",
                          "tipoPessoa",
                          "ativo",
                          "criadoEm",
                          "atualizadoEm"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "page",
                        "limit",
                        "total",
                        "totalPages"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/veiculos/": {
      "post": {
        "summary": "Cadastrar veículo",
        "tags": [
          "Veículos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clienteId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "placa": {
                    "type": "string"
                  },
                  "marca": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "modelo": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "ano": {
                    "type": "integer",
                    "minimum": 1900,
                    "maximum": 2027
                  },
                  "cor": {
                    "type": "string",
                    "maxLength": 50
                  }
                },
                "required": [
                  "clienteId",
                  "placa",
                  "marca",
                  "modelo",
                  "ano"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "placa": {
                      "type": "string"
                    },
                    "marca": {
                      "type": "string"
                    },
                    "modelo": {
                      "type": "string"
                    },
                    "ano": {
                      "type": "integer"
                    },
                    "cor": {
                      "type": "string",
                      "nullable": true
                    },
                    "ativo": {
                      "type": "boolean"
                    },
                    "clienteId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "criadoEm": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "atualizadoEm": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "placa",
                    "marca",
                    "modelo",
                    "ano",
                    "cor",
                    "ativo",
                    "clienteId",
                    "criadoEm",
                    "atualizadoEm"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar veículos",
        "tags": [
          "Veículos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "clienteId",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "placa": {
                        "type": "string"
                      },
                      "marca": {
                        "type": "string"
                      },
                      "modelo": {
                        "type": "string"
                      },
                      "ano": {
                        "type": "integer"
                      },
                      "cor": {
                        "type": "string",
                        "nullable": true
                      },
                      "ativo": {
                        "type": "boolean"
                      },
                      "clienteId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "criadoEm": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "atualizadoEm": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "id",
                      "placa",
                      "marca",
                      "modelo",
                      "ano",
                      "cor",
                      "ativo",
                      "clienteId",
                      "criadoEm",
                      "atualizadoEm"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}